+++ zcash_client_backend/src/keys.rs
@@ -1,5 +1,4 @@
 //! Helper functions for managing light client key material.
-use orchard;
 use zcash_address::unified::{self, Container, Encoding};
 use zcash_primitives::{
     consensus,
@@ -84,6 +83,7 @@ fn to_transparent_child_index(j: DiversifierIndex) -> Option<u32> {
 #[derive(Debug)]
 #[doc(hidden)]
 pub enum DerivationError {
+    #[cfg(feature = "orchard")]
     Orchard(orchard::zip32::Error),
     #[cfg(feature = "transparent-inputs")]
     Transparent(hdwallet::error::Error),
@@ -144,6 +144,7 @@ pub struct UnifiedSpendingKey {
     #[cfg(feature = "transparent-inputs")]
     transparent: legacy::AccountPrivKey,
     sapling: sapling::ExtendedSpendingKey,
+    #[cfg(feature = "orchard")]
     orchard: orchard::keys::SpendingKey,
 }
 
@@ -158,6 +159,7 @@ impl UnifiedSpendingKey {
             panic!("ZIP 32 seeds MUST be at least 32 bytes");
         }
 
+        #[cfg(feature = "orchard")]
         let orchard =
             orchard::keys::SpendingKey::from_zip32_seed(seed, params.coin_type(), account.into())
                 .map_err(DerivationError::Orchard)?;
@@ -170,6 +172,7 @@ impl UnifiedSpendingKey {
             #[cfg(feature = "transparent-inputs")]
             transparent,
             sapling: sapling::spending_key(seed, params.coin_type(), account),
+            #[cfg(feature = "orchard")]
             orchard,
         })
     }
@@ -179,6 +182,7 @@ impl UnifiedSpendingKey {
             #[cfg(feature = "transparent-inputs")]
             transparent: Some(self.transparent.to_account_pubkey()),
             sapling: Some(self.sapling.to_diversifiable_full_viewing_key()),
+            #[cfg(feature = "orchard")]
             orchard: Some((&self.orchard).into()),
             unknown: vec![],
         }
@@ -197,6 +201,7 @@ impl UnifiedSpendingKey {
     }
 
     /// Returns the Orchard spending key component of this unified spending key.
+    #[cfg(feature = "orchard")]
     pub fn orchard(&self) -> &orchard::keys::SpendingKey {
         &self.orchard
     }
@@ -215,13 +220,15 @@ impl UnifiedSpendingKey {
         let mut result = vec![];
         result.write_u32::<LittleEndian>(era.id()).unwrap();
 
-        // orchard
-        let orchard_key = self.orchard();
-        CompactSize::write(&mut result, usize::try_from(Typecode::Orchard).unwrap()).unwrap();
+        #[cfg(feature = "orchard")]
+        {
+            let orchard_key = self.orchard();
+            CompactSize::write(&mut result, usize::try_from(Typecode::Orchard).unwrap()).unwrap();
 
-        let orchard_key_bytes = orchard_key.to_bytes();
-        CompactSize::write(&mut result, orchard_key_bytes.len()).unwrap();
-        result.write_all(orchard_key_bytes).unwrap();
+            let orchard_key_bytes = orchard_key.to_bytes();
+            CompactSize::write(&mut result, orchard_key_bytes.len()).unwrap();
+            result.write_all(orchard_key_bytes).unwrap();
+        }
 
         // sapling
         let sapling_key = self.sapling();
@@ -261,6 +268,7 @@ impl UnifiedSpendingKey {
             return Err(DecodingError::EraMismatch(decoded_era));
         }
 
+        #[cfg(feature = "orchard")]
         let mut orchard = None;
         let mut sapling = None;
         #[cfg(feature = "transparent-inputs")]
@@ -283,12 +291,16 @@ impl UnifiedSpendingKey {
                     source
                         .read_exact(&mut key)
                         .map_err(|_| DecodingError::InsufficientData(Typecode::Orchard))?;
-                    orchard = Some(
-                        Option::<orchard::keys::SpendingKey>::from(
-                            orchard::keys::SpendingKey::from_bytes(key),
-                        )
-                        .ok_or(DecodingError::KeyDataInvalid(Typecode::Orchard))?,
-                    );
+
+                    #[cfg(feature = "orchard")]
+                    {
+                        orchard = Some(
+                            Option::<orchard::keys::SpendingKey>::from(
+                                orchard::keys::SpendingKey::from_bytes(key),
+                            )
+                            .ok_or(DecodingError::KeyDataInvalid(Typecode::Orchard))?,
+                        );
+                    }
                 }
                 Typecode::Sapling => {
                     if len != 169 {
@@ -324,13 +336,19 @@ impl UnifiedSpendingKey {
                 }
             }
 
+            #[cfg(feature = "orchard")]
+            let has_orchard = orchard.is_some();
+            #[cfg(not(feature = "orchard"))]
+            let has_orchard = true;
+
             #[cfg(feature = "transparent-inputs")]
             let has_transparent = transparent.is_some();
             #[cfg(not(feature = "transparent-inputs"))]
             let has_transparent = true;
 
-            if orchard.is_some() && sapling.is_some() && has_transparent {
+            if has_orchard && sapling.is_some() && has_transparent {
                 return Ok(UnifiedSpendingKey {
+                    #[cfg(feature = "orchard")]
                     orchard: orchard.unwrap(),
                     sapling: sapling.unwrap(),
                     #[cfg(feature = "transparent-inputs")]
@@ -362,6 +380,7 @@ pub struct UnifiedFullViewingKey {
     #[cfg(feature = "transparent-inputs")]
     transparent: Option<legacy::AccountPubKey>,
     sapling: Option<sapling::DiversifiableFullViewingKey>,
+    #[cfg(feature = "orchard")]
     orchard: Option<orchard::keys::FullViewingKey>,
     unknown: Vec<(u32, Vec<u8>)>,
 }
@@ -372,7 +391,7 @@ impl UnifiedFullViewingKey {
     pub fn new(
         #[cfg(feature = "transparent-inputs")] transparent: Option<legacy::AccountPubKey>,
         sapling: Option<sapling::DiversifiableFullViewingKey>,
-        orchard: Option<orchard::keys::FullViewingKey>,
+        #[cfg(feature = "orchard")] orchard: Option<orchard::keys::FullViewingKey>,
     ) -> Option<UnifiedFullViewingKey> {
         if sapling.is_none() {
             None
@@ -381,6 +400,7 @@ impl UnifiedFullViewingKey {
                 #[cfg(feature = "transparent-inputs")]
                 transparent,
                 sapling,
+                #[cfg(feature = "orchard")]
                 orchard,
                 // We don't allow constructing new UFVKs with unknown items, but we store
                 // this to allow parsing such UFVKs.
@@ -402,6 +422,7 @@ impl UnifiedFullViewingKey {
             ));
         }
 
+        #[cfg(feature = "orchard")]
         let mut orchard = None;
         let mut sapling = None;
         #[cfg(feature = "transparent-inputs")]
@@ -413,6 +434,7 @@ impl UnifiedFullViewingKey {
             .items_as_parsed()
             .iter()
             .filter_map(|receiver| match receiver {
+                #[cfg(feature = "orchard")]
                 unified::Fvk::Orchard(data) => orchard::keys::FullViewingKey::from_bytes(data)
                     .ok_or("Invalid Orchard FVK in Unified FVK")
                     .map(|addr| {
@@ -420,6 +442,10 @@ impl UnifiedFullViewingKey {
                         None
                     })
                     .transpose(),
+                #[cfg(not(feature = "orchard"))]
+                unified::Fvk::Orchard(data) => {
+                    Some(Ok((unified::Typecode::Orchard.into(), data.to_vec())))
+                }
                 unified::Fvk::Sapling(data) => {
                     sapling::DiversifiableFullViewingKey::from_bytes(data)
                         .ok_or("Invalid Sapling FVK in Unified FVK")
@@ -449,6 +475,7 @@ impl UnifiedFullViewingKey {
             #[cfg(feature = "transparent-inputs")]
             transparent,
             sapling,
+            #[cfg(feature = "orchard")]
             orchard,
             unknown,
         })
@@ -457,12 +484,6 @@ impl UnifiedFullViewingKey {
     /// Returns the string encoding of this `UnifiedFullViewingKey` for the given network.
     pub fn encode<P: consensus::Parameters>(&self, params: &P) -> String {
         let items = std::iter::empty()
-            .chain(
-                self.orchard
-                    .as_ref()
-                    .map(|fvk| fvk.to_bytes())
-                    .map(unified::Fvk::Orchard),
-            )
             .chain(
                 self.sapling
                     .as_ref()
@@ -477,6 +498,13 @@ impl UnifiedFullViewingKey {
                         data: data.clone(),
                     }),
             );
+        #[cfg(feature = "orchard")]
+        let items = items.chain(
+            self.orchard
+                .as_ref()
+                .map(|fvk| fvk.to_bytes())
+                .map(unified::Fvk::Orchard),
+        );
         #[cfg(feature = "transparent-inputs")]
         let items = items.chain(
             self.transparent
@@ -503,6 +531,7 @@ impl UnifiedFullViewingKey {
     }
 
     /// Returns the Orchard full viewing key component of this unified key.
+    #[cfg(feature = "orchard")]
     pub fn orchard(&self) -> Option<&orchard::keys::FullViewingKey> {
         self.orchard.as_ref()
     }
@@ -537,7 +566,12 @@ impl UnifiedFullViewingKey {
         #[cfg(not(feature = "transparent-inputs"))]
         let transparent = None;
 
-        UnifiedAddress::from_receivers(None, sapling, transparent)
+        UnifiedAddress::from_receivers(
+            #[cfg(feature = "orchard")]
+            None,
+            sapling,
+            transparent,
+        )
     }
 
     /// Searches the diversifier space starting at diversifier index `j` for one which will
@@ -658,6 +692,7 @@ mod tests {
     fn ufvk_round_trip() {
         let account = AccountId::ZERO;
 
+        #[cfg(feature = "orchard")]
         let orchard = {
             let sk = orchard::keys::SpendingKey::from_zip32_seed(&[0; 32], 0, 0).unwrap();
             Some(orchard::keys::FullViewingKey::from(&sk))
@@ -678,19 +713,28 @@ mod tests {
             #[cfg(feature = "transparent-inputs")]
             transparent,
             sapling,
+            #[cfg(feature = "orchard")]
             orchard,
         )
         .unwrap();
 
         let encoded = ufvk.encode(&MAIN_NETWORK);
 
-        // test encoded form against known values
+        // Test encoded form against known values; these test vectors contain Orchard receivers
+        // that will be treated as unknown if the `orchard` feature is not enabled.
         let encoded_with_t = "uview1tg6rpjgju2s2j37gkgjq79qrh5lvzr6e0ed3n4sf4hu5qd35vmsh7avl80xa6mx7ryqce9hztwaqwrdthetpy4pc0kce25x453hwcmax02p80pg5savlg865sft9reat07c5vlactr6l2pxtlqtqunt2j9gmvr8spcuzf07af80h5qmut38h0gvcfa9k4rwujacwwca9vu8jev7wq6c725huv8qjmhss3hdj2vh8cfxhpqcm2qzc34msyrfxk5u6dqttt4vv2mr0aajreww5yufpk0gn4xkfm888467k7v6fmw7syqq6cceu078yw8xja502jxr0jgum43lhvpzmf7eu5dmnn6cr6f7p43yw8znzgxg598mllewnx076hljlvynhzwn5es94yrv65tdg3utuz2u3sras0wfcq4adxwdvlk387d22g3q98t5z74quw2fa4wed32escx8dwh4mw35t4jwf35xyfxnu83mk5s4kw2glkgsshmxk";
         let _encoded_no_t = "uview12z384wdq76ceewlsu0esk7d97qnd23v2qnvhujxtcf2lsq8g4hwzpx44fwxssnm5tg8skyh4tnc8gydwxefnnm0hd0a6c6etmj0pp9jqkdsllkr70u8gpf7ndsfqcjlqn6dec3faumzqlqcmtjf8vp92h7kj38ph2786zx30hq2wru8ae3excdwc8w0z3t9fuw7mt7xy5sn6s4e45kwm0cjp70wytnensgdnev286t3vew3yuwt2hcz865y037k30e428dvgne37xvyeal2vu8yjnznphf9t2rw3gdp0hk5zwq00ws8f3l3j5n3qkqgsyzrwx4qzmgq0xwwk4vz2r6vtsykgz089jncvycmem3535zjwvvtvjw8v98y0d5ydwte575gjm7a7k";
-        #[cfg(feature = "transparent-inputs")]
-        assert_eq!(encoded, encoded_with_t);
-        #[cfg(not(feature = "transparent-inputs"))]
-        assert_eq!(encoded, _encoded_no_t);
+
+        // We test the full roundtrip only with the `orchard` feature enabled, because we will
+        // not generate the `orchard` part of the encoding if the UFVK does not have an Orchard
+        // part.
+        #[cfg(feature = "orchard")]
+        {
+            #[cfg(feature = "transparent-inputs")]
+            assert_eq!(encoded, encoded_with_t);
+            #[cfg(not(feature = "transparent-inputs"))]
+            assert_eq!(encoded, _encoded_no_t);
+        }
 
         let decoded = UnifiedFullViewingKey::decode(&MAIN_NETWORK, &encoded).unwrap();
         let reencoded = decoded.encode(&MAIN_NETWORK);
@@ -705,6 +749,7 @@ mod tests {
             decoded.sapling.map(|s| s.to_bytes()),
             ufvk.sapling.map(|s| s.to_bytes()),
         );
+        #[cfg(feature = "orchard")]
         assert_eq!(
             decoded.orchard.map(|o| o.to_bytes()),
             ufvk.orchard.map(|o| o.to_bytes()),
@@ -716,7 +761,12 @@ mod tests {
             decoded_with_t.transparent.map(|t| t.serialize()),
             ufvk.transparent.as_ref().map(|t| t.serialize()),
         );
-        #[cfg(not(feature = "transparent-inputs"))]
+
+        #[cfg(all(not(feature = "orchard"), not(feature = "transparent-inputs")))]
+        assert_eq!(decoded_with_t.unknown.len(), 2);
+        #[cfg(all(feature = "transparent-inputs", not(feature = "orchard")))]
+        assert_eq!(decoded_with_t.unknown.len(), 1);
+        #[cfg(all(feature = "orchard", not(feature = "transparent-inputs")))]
         assert_eq!(decoded_with_t.unknown.len(), 1);
     }
 
@@ -771,14 +821,20 @@ mod tests {
         #[cfg(feature = "unstable")]
         fn prop_usk_roundtrip(usk in arb_unified_spending_key(Network::MainNetwork)) {
             let encoded = usk.to_bytes(Era::Orchard);
+
             #[cfg(not(feature = "transparent-inputs"))]
             assert_eq!(encoded.len(), 4 + 2 + 32 + 2 + 169);
             #[cfg(feature = "transparent-inputs")]
             assert_eq!(encoded.len(), 4 + 2 + 32 + 2 + 169 + 2 + 64);
+
             let decoded = UnifiedSpendingKey::from_bytes(Era::Orchard, &encoded);
             let decoded = decoded.unwrap_or_else(|e| panic!("Error decoding USK: {:?}", e));
+
+            #[cfg(feature = "orchard")]
             assert!(bool::from(decoded.orchard().ct_eq(usk.orchard())));
+
             assert_eq!(decoded.sapling(), usk.sapling());
+
             #[cfg(feature = "transparent-inputs")]
             assert_eq!(decoded.transparent().to_bytes(), usk.transparent().to_bytes());
         }